StorageHandler & SecretsHandler refactor
There were a couple of problems with the original design that are handled
(pun intended) here:
1. StorageHandler was merely a struct that held other object to couple them together.
It had no methods of its own and was used only to trigger methods of the objects it held.
2. The word "handler" was very ambiguous

This refactor removes the StorageHandler (and moves its config to Secrets) and now we have only an AccessToken and a SecretsHandler.
The SecretsHandler was refactored to ProvideConjurSecrets to improve readability of this
struct.

The last change was to restructure the code to have better hierarchy.