Added UTs for RetrieveRequiredK8sSecrets & Run methods
In order to write UTs for these methods there was a need to refactor the code
a bit. Some methods were changed so that instead of initializing a struct,
it receives an interface as an input. This way, app code can call the method
with the initialized struct and UT code can call the method with a Mock that
implements the same interface. For this cause, some interfaces were created
in this commit. In addition, the k8s_secrets_client now only communicates with
k8s. All the logic including which secrets should be retrieved or patched was
moved to ProvideConjurSecretsToK8sSecrets.

To improve readability of UTs, every mocked object is located in its own file.