Improve docker image versioning process
Until this commit, in each green `master` build we pushed 2 images to our
docker registry:
- a `latest` image
- a fixed version image

This process is not ideal as we replace the stable image (i.e `0.2.0`)
after each green `master` build. In this commit, we push a stable fixed image
only when we release a version in git.