Remove `openshift` annotation from conjur hosts in the policy
The `openshift: true` & `kubernetes: true` annotations enable a feature
in the UI where if this annotation exists the UI displays a logo for
the host.

We don't use this feature in this repo and because we load the same
policies for kubernetes & openshift it may be misleading for someone
who goes over the policy files.