Provide complex secrets (#79)
* Add support for complex secrets

The secrets-provider had as issue with providing complex secrets to k8s.
For example, the secret `{"auths":{"someurl":{"auth":"sometoken="}}}` failed to be provided from Conjur.

Changing the method from PATCH to UPDATE while pushing the secret into k8s solved the issue

Give service-account `update` permissions for tests

We now provide secrets to k8s using the `UPDATE` method instead of
`PATCH`. This commit updates the service-account used in our tests to have
`update` permissions instead of `patch`.

Signed-off-by: Samuel Prette <samuel.prette@clearstream.com>

* Test that complex secrets are provided to k8s

Instead of testing that the secret `supersecret` is provided successfully to k8s, we now test more complex secrets, such as JSON Objects and SSH Keys. We don't test also simple secrets as if complex one are provided then simple one will be as well.

We run the tests in different files for 2 reasons:
  1. In case of failure it is clear which complex secret failed.
  2. We need to tear down and re-create the environment in order to test
     these cases so the secrets-provider pod is re-created.

* Improve imports formatting

Signed-off-by: Samuel Prette <samuel.prette@clearstream.com>

* Fix tests use from `patch` to `update`

We changed the way we provide secrets into k8s from `patch`
to `update` so the tests should be updated accordingly.

* Add entry to changelog about complex secrets fix

* Fix indentation in Jenkinsfile

* Test that providing secrets from Conjur doesn't damage k8s secret

We want to verify that the K8s secret's keys are not damaged in the
secrets-provider process.

For example, if the k8s secret has the keys `conjur-map` and `another-key`
then the value of `abother-key` stays intact after we provide the values
of `conjur-map`

* Update CHANGELOG entry to tell the user to use `update` permissions

Co-authored-by: Samuel Prette <samuel.prette@clearstream.com>