Use k8s apply instead of create in loops
Using create in a loop that creates multiple resources can lead to
deadlock. For example:
    Trying to create resources a and b, on attempt 1 resource a
    is created, on attempt two, a fails to create because it already
    exists, and b is created. Now both resources have been created
    but neither attempt at creating them exited with success, so
    the loop will continue trying to create them. All subsequent
    attempts to create them will fail as they already exist.

Using apply instead of create resolves this deadlock as no changes
need to be made to an existing resource when re-applying the same
configuration.

Fixing this because it caused my previous commit in conjuirnc/ops#583
to fail.

Related: conjurinc/ops#583