Remove unused ByteSlicePrintf util (#115)
* Remove unused ByteSlicePrintf util

When we changed the K8s secret update method from `patch` to `update`
we stopped using the util we built for it but didn't remove the util
itself.

* Change test file name to better tell its use-case

Change the test file from "no patch permissions" to "no update permissions"
to better tell the use-case. We previously changed the required permissions
to `get, update` instead of `get, patch`. This test verifies the expected
behaviour when the `update` permission in missing.