Add UTs for config package (#118)
* Rename getStoreType method to better fits its logic

The method getStoreType receives the value of the env var
 `SECRETS_DESTINATION` and checks if it is valid. If so, it returns
 the same value. If not, it raises an error. Eventually it validates
 the value of `SECRETS_DESTINATION` so it is best to name it that way
 and change the logic to fit its purpose.

* Add UTs for config package

Previously we had only one UT for this package and it tested a
private method (which we shouldn't do). This commit removed the
existing UT and added instead UTs for the public method `NewFromEnv`.

It verifies all of its use-cases and has 100% coverage.