Fix release build to push to DockerHub
When releasing, we were facing a permissions, request to resource denied error. The problem originates from `docker login` inside a container and mapping of docker configuration inside a container. The mapping that this PR removes maps the host docker config into the container. This affects the file on the host and the Jenkins user couldn't access the file outside of the container to push the image.