Add bash util to get pod name
The method to get the pod name was sometimes finding no pods and not retrying.
In this commit, we create a new bash method get_pod_name in utils.sh that
searches for the pod name using the jsonpath, which errors if there are no
matching pods so that the wait_for_it script will retry as needed. The method
also errors if after all of the retries, there is no matching pod. On error,
it prints the namespace pod description.

It also updates the test cases to use this new util method when checking for
the pod name.