Adds Helm chart option to set Secrets Provider Job name
Currently, the Secrets Provider Helm chart deploys a Secrets Provider
Kubernetes Job with a name that is (unconditionally) set to the Helm Release
name.

In the case where the Secrets Provider Helm chart is included as a
[Helm dependency](https://helm.sh/docs/helm/helm_dependency) to another
(parent) Helm chart, the Secrets Provider Helm chart inherits its Release
name from its parent chart. This currently results in the Secrets Provider
Job name being set to the Release name of the **parent** Helm chart, which
is a bit confusing.

This change adds a chart value to allow the Secrets Provider Kubernetes Job
to be explicitly set, if desired. If this chart value is not set explicitly,
then the Secrets Provider Kubernetes Job name will default to the chart's
Release name.

This change also includes a couple of Helm unittest test cases to test
the precedence ordering for the SP Job name.