Refactor SP main so that trace deferred calls will work
The SP main.go had a lot of scattered instances of calls to os.Exit().
Since os.Exit() doesn't call an functions that are registered as
deferred, we weren't getting traces/spans properly closed and reported
to a collector.

Added some calls to register Spans as errored, with updated status so
that we can use traces for debugging/troubleshooting.

Also, split up the interfaces in the pkg/trace package so that
TracerProvider, Tracer, and Span interfaces are implemented in their own
source files.

Also, started adding support a "Noop" TracerProvider... but this
currently doesn't work (SP seems to crash early, but it's crashing
silently).