Build and push images on version tags
The Jenkinsfile includes a when conditional intended to prevent the pipeline
from building and releasing when only markdown files have been changed. It
unfortunately also prevents a build and release when a version tag has been
created since no files are changed from the main branch. We fix this by also
requiring build and release on tags with a naming scheme of v*.