List of paths to retrieve from Conjur should be a set
Currently, in Push to File mode, when a list of secret paths for secrets
that need to be retrieved from Conjur is compiled across all Secret Groups,
there is no logic to eliminate duplicate paths. Apparently, Conjur is
not tolerant of duplicate file paths, since this causes retrieval
failures. The fix is to ensure that the list of secret paths to retrieve
from Conjur is a mathematical set (i.e. contains only unique elements).

Also fixed another bug whereby if no secrets need to be retrieved from Conjur,
we still make a secrets retrieval request to Conjur with an empty list
of secret paths, and Conjur is not tolerant of this.