Don't overwrite secret file if atomic temp file not written
When secret rotation is enabled in Push-to-File mode, and the Secrets
Provider (periodically) checks for changes in secret files and finds no change
in content, it correctly refrains from writing rendered file content to the
temp file, as expected.

However, even though no content has been written to the temp file, there
is a deferred call to the atomic file writer's Close() method that
unconditionally renames the temp file to the destination secret file,
effectively deleting the secret file content (i.e. overwriting the file
with an empty file).

This change adds a check in the Close() method so that we don't
overwrite the secret file if the associated temp file has not been
written e.g. because there is no change in secret file content.

This change also adds another check so that we don't attempt to delete
the temp file if it has already been renamed to the target secret file
name.